Extract PackageProcessor into separate file
- Adjust tests to interact with this class via its public API instead of
manipulating internal state of UpdateRelease.
- Use class method to signal ephemeral nature of this object
