+++ src/bosh-director/lib/bosh/director/jobs/update_release.rb
@@ -201,142 +201,13 @@ module Bosh::Director
         end
       end
 
-      class PackageProcessor
-        def initialize(
-          release_dir,
-          compiled_release,
-          release_version_model,
-          release_model,
-          name,
-          version,
-          manifest_packages,
-          logger,
-          update_release
-        )
-
-          @release_dir = release_dir
-          @compiled_release = compiled_release
-          @release_version_model = release_version_model
-          @release_model = release_model
-          @name = name
-          @version = version
-          @manifest_packages = manifest_packages
-          @logger = logger
-          @update_release = update_release
-
-          @new_packages = []
-          @existing_packages = []
-          @registered_packages = []
-        end
-
-        attr_reader :logger, :release_dir, :manifest_packages, :update_release
-
-        def process
-          logger.info('Checking for new packages in release')
-
-          manifest_packages.each do |package_meta|
-            package_meta['compiled_package_sha1'] = package_meta['sha1']
-
-            validate_package_fingerprint!(package_meta)
-
-            packages = Models::Package.where(fingerprint: package_meta['fingerprint']).order_by(:id).all
-
-            if packages.empty?
-              @new_packages << package_meta
-              next
-            end
-
-            existing_package = packages.find do |package|
-              package.release_id == @release_model.id &&
-                package.name == package_meta['name'] &&
-                package.version == package_meta['version']
-            end
-
-            if existing_package
-              use_existing_package(existing_package, packages, package_meta)
-            else
-              reuse_package_matching_fingerprint(packages, package_meta)
-            end
-          end
-
-          persist_packages
-        end
-
-        private
-
-        def validate_package_fingerprint!(package_meta)
-          # Checking whether we might have the same bits somewhere (in any release, not just the one being uploaded)
-          @release_version_model.packages.select { |pv| pv.name == package_meta['name'] }.each do |package|
-            if package.fingerprint !=  package_meta['fingerprint']
-              raise ReleaseInvalidPackage, "package '#{package_meta['name']}' had different fingerprint in previously uploaded release '#{@name}/#{@version}'"
-            end
-          end
-        end
-
-        def use_existing_package(existing_package, packages, package_meta)
-          if existing_package.blobstore_id.nil?
-            packages.each do |package|
-              next if package.blobstore_id.nil?
-
-              package_meta['blobstore_id'] = package.blobstore_id
-              package_meta['sha1'] = package.sha1
-              break
-            end
-          end
-
-          if existing_package.release_versions.include?(@release_version_model)
-            @registered_packages << [existing_package, package_meta]
-          else
-            @existing_packages << [existing_package, package_meta]
-          end
-        end
-
-        def reuse_package_matching_fingerprint(packages, package_meta)
-          # We found a package with the same fingerprint but different
-          # (release, name, version) tuple, so we need to make a copy
-          # of the package blob and create a new db entry for it
-          packages.each do |package|
-            next if package.blobstore_id.nil?
-
-            package_meta['blobstore_id'] = package.blobstore_id
-            package_meta['sha1'] = package.sha1
-            break
-          end
-
-          @new_packages << package_meta
-        end
-
-        def persist_packages
-          created_package_refs = update_release.create_packages(@new_packages, release_dir)
-
-          existing_package_refs = update_release.use_existing_packages(@existing_packages, release_dir)
-
-          if @compiled_release
-            all_package_refs = Array(created_package_refs) | Array(existing_package_refs) | registered_package_refs
-            update_release.create_compiled_packages(all_package_refs, release_dir)
-            return
-          end
-
-          update_release.backfill_source_for_packages(@registered_packages, release_dir)
-        end
-
-        def registered_package_refs
-          @registered_packages.map do |pkg, pkg_meta|
-            {
-              package: pkg,
-              package_meta: pkg_meta,
-            }
-          end
-        end
-      end
-
       # Finds all package definitions in the manifest and sorts them into two
       # buckets: new and existing packages, then creates new packages and points
       # current release version to the existing packages.
       # @param [String] release_dir local path to the unpacked release
       # @return [void]
       def process_packages(release_dir)
-        PackageProcessor.new(
+        PackageProcessor.process(
           release_dir,
           @compiled_release,
           @release_version_model,
@@ -346,7 +217,7 @@ module Bosh::Director
           manifest_packages,
           logger,
           self,
-        ).process
+        )
       end
 
       # @return [boolean] true if sources were added to at least one package; false if the call had no effect.
