+++ src/bosh-director/spec/unit/jobs/update_release_spec.rb
@@ -1381,336 +1381,6 @@ module Bosh::Director
       end
     end
 
-    describe 'process_packages' do
-      subject(:update_release) { Jobs::UpdateRelease.new(release_dir) }
-      let(:release_dir) { double(:release_dir) }
-
-      before do
-        allow(update_release).to receive(:manifest_packages).and_return []
-      end
-
-      context 'if it is compiled' do
-        before do
-          update_release.instance_variable_set(:@compiled_release, true)
-        end
-
-        it 'creates compiled packages' do
-          expect(update_release).to receive(:create_compiled_packages).with([], release_dir)
-          update_release.process_packages(release_dir)
-        end
-      end
-
-      context 'if it is not compiled' do
-        before do
-          update_release.instance_variable_set(:@compiled_release, false)
-        end
-
-        it 'backfills source for packages' do
-          expect(update_release).to receive(:backfill_source_for_packages).with([], release_dir)
-          update_release.process_packages(release_dir)
-        end
-      end
-
-      context 'if a package fingerprint changes' do
-        before do
-          allow(update_release).to receive(:manifest_packages).and_return [
-            { 'name' => 'package_a', 'fingerprint' => 'b' },
-          ]
-
-          allow(release_version_model).to receive(:packages).and_return [
-            instance_double(Models::Package, name: 'package_a', fingerprint: 'a'),
-          ]
-
-          update_release.instance_variable_set(:@release_version_model, release_version_model)
-          update_release.instance_variable_set(:@name, 'release-name')
-          update_release.instance_variable_set(:@version, 'release-version')
-        end
-
-        let(:release_version_model) { instance_double(Models::ReleaseVersion) }
-
-        it 'raises an exception' do
-          expect do
-            update_release.process_packages(release_dir)
-          end.to raise_error(
-            "package 'package_a' had different fingerprint in previously uploaded release 'release-name/release-version'",
-          )
-        end
-      end
-
-      context 'if there is no existing package with the same fingerprint' do
-        before do
-          allow(update_release).to receive(:manifest_packages).and_return [
-            new_package_metadata,
-          ]
-
-          allow(release_version_model).to receive(:packages).and_return []
-          allow(update_release).to receive(:create_packages).and_return(create_compiled_packages_response)
-
-          update_release.instance_variable_set(:@release_version_model, release_version_model)
-        end
-
-        let(:new_package_metadata) do
-          { 'name' => 'package-a', 'fingerprint' => 'print-a', 'sha1' => 'sha-a' }
-        end
-
-        let(:create_compiled_packages_response) { [double(:create_compiled_packages_response)] }
-
-        let(:release_version_model) { instance_double(Models::ReleaseVersion) }
-
-        it 'treats it as a new package' do
-          expect(update_release).to receive(:create_packages).with(
-            [new_package_metadata.merge('compiled_package_sha1' => new_package_metadata['sha1'])],
-            release_dir,
-          )
-
-          update_release.process_packages(release_dir)
-        end
-
-        context 'if it is compiled' do
-          before do
-            update_release.instance_variable_set(:@compiled_release, true)
-          end
-
-          it 'creates compiled packages' do
-            expect(update_release).to receive(:create_compiled_packages).with(create_compiled_packages_response, release_dir)
-            update_release.process_packages(release_dir)
-          end
-        end
-
-        context 'if it is not compiled' do
-          before do
-            update_release.instance_variable_set(:@compiled_release, false)
-          end
-
-          it 'backfills source for packages, ignoring the new package' do
-            expect(update_release).to receive(:backfill_source_for_packages).with([], release_dir)
-            update_release.process_packages(release_dir)
-          end
-        end
-      end
-
-      context 'if there exists a package that has a matching fingerprint' do
-        before do
-          release_model_id = 1
-
-          allow(update_release).to receive(:manifest_packages).and_return [
-            new_package_metadata,
-          ]
-
-          update_release.instance_variable_set(:@release_model, double(id: release_model_id))
-          update_release.instance_variable_set(:@release_version_model, release_version_model)
-          allow(update_release).to receive(:create_packages)
-        end
-
-        let!(:release_version_model) do
-          release.add_version(Models::ReleaseVersion.make)
-        end
-        let(:new_package_metadata) do
-          {
-            'name' => 'package-a',
-            'version' => 'package-version',
-            'fingerprint' => package_fingerprint,
-            'sha1' => 'sha-a',
-          }
-        end
-
-        let(:release) { Models::Release.make(name: 'appcloud') }
-        let!(:package_model) do
-          package_id = Models::Package.insert(
-            name: package_name,
-            fingerprint: package_fingerprint,
-            sha1: model_sha,
-            version: package_version,
-            blobstore_id: blobstore_id,
-            dependency_set_json: '{}',
-            release_id: release.id,
-          )
-          Models::Package.where(id: package_id).first
-        end
-
-        let(:package_name) { 'package-a' }
-        let(:package_version) { 'package-version' }
-        let(:model_sha) { 'model-sha-a' }
-        let(:package_fingerprint) { 'a' }
-        let(:blobstore_id) { '1234' }
-
-        context 'and is preexisting' do
-          context 'and is not registered already' do
-            context 'and has a non-nil blobstore id' do
-              it 'does not change the package metadata' do
-                expect(update_release).to receive(:use_existing_packages).with(
-                  [[
-                    package_model,
-                    new_package_metadata.merge('compiled_package_sha1' => new_package_metadata['sha1']),
-                  ]],
-                  release_dir,
-                )
-
-                update_release.process_packages(release_dir)
-              end
-            end
-
-            context 'but has a nil blobstore id' do
-              let(:blobstore_id) { nil }
-
-              it 'copies the blobstore id from another package with the same fingerprint into the package metadata' do
-                another_blobstore_id = '2345'
-                another_model_sha1 = 'model-sha-b'
-                Models::Package.insert(
-                  name: package_name,
-                  fingerprint: package_fingerprint,
-                  sha1: another_model_sha1,
-                  version: package_version,
-                  blobstore_id: another_blobstore_id,
-                  dependency_set_json: '{}',
-                  release_id: release.id,
-                )
-
-                expect(update_release).to receive(:use_existing_packages).with(
-                  [[
-                    package_model,
-                    new_package_metadata.merge(
-                      'compiled_package_sha1' => new_package_metadata['sha1'],
-                      'blobstore_id' => another_blobstore_id,
-                      'sha1' => another_model_sha1,
-                    ),
-                  ]],
-                  release_dir,
-                )
-
-                update_release.process_packages(release_dir)
-              end
-            end
-          end
-
-          context 'and is registered already' do
-            before do
-              release_version_model.add_package(package_model)
-            end
-
-            context 'and is compiled' do
-              before do
-                update_release.instance_variable_set(:@compiled_release, true)
-              end
-
-              it 'reshapes the existing packages to input into create_compiled_packages' do
-                expect(update_release).to receive(:create_compiled_packages).with(
-                  [{
-                    package: package_model,
-                    package_meta: new_package_metadata,
-                  }],
-                  release_dir,
-                )
-
-                update_release.process_packages(release_dir)
-              end
-            end
-
-            context 'and is not compiled' do
-              before do
-                update_release.instance_variable_set(:@compiled_release, false)
-              end
-
-              it 'backfills source for packages' do
-                expect(update_release).to receive(:backfill_source_for_packages).with(
-                  [[
-                    package_model,
-                    new_package_metadata.merge('compiled_package_sha1' => new_package_metadata['sha1']),
-                  ]],
-                  release_dir,
-                )
-
-                update_release.process_packages(release_dir)
-              end
-            end
-          end
-        end
-
-        context 'and is not preexisting' do
-          context 'and has a nil blobstore id' do
-            before do
-              # Need to make one of release, name, or version disagree
-              update_release.instance_variable_set(:@release_model, double(id: 2))
-            end
-
-            let(:blobstore_id) { nil }
-
-            it 'does not reuse the blob' do
-              expect(update_release).to receive(:create_packages).with(
-                [hash_excluding('blobstore_id' => blobstore_id, 'sha1' => model_sha)],
-                release_dir,
-              )
-
-              update_release.process_packages(release_dir)
-            end
-          end
-
-          context 'and has a blobstore id' do
-            let(:blobstore_id) { '1234' }
-
-            context 'but does not match release id' do
-              before do
-                update_release.instance_variable_set(:@release_model, double(id: 2))
-              end
-
-              it 'copies metadata from the existing package' do
-                expect(update_release).to receive(:create_packages).with(
-                  [
-                    new_package_metadata.merge(
-                      'compiled_package_sha1' => new_package_metadata['sha1'],
-                      'blobstore_id' => blobstore_id,
-                      'sha1' => model_sha,
-                    ),
-                  ],
-                  release_dir,
-                )
-
-                update_release.process_packages(release_dir)
-              end
-            end
-
-            context 'but does not match name' do
-              let(:package_name) { 'package-b' }
-
-              it 'copies metadata from the existing package' do
-                expect(update_release).to receive(:create_packages).with(
-                  [
-                    new_package_metadata.merge(
-                      'compiled_package_sha1' => new_package_metadata['sha1'],
-                      'blobstore_id' => blobstore_id,
-                      'sha1' => model_sha,
-                    ),
-                  ],
-                  release_dir,
-                )
-
-                update_release.process_packages(release_dir)
-              end
-            end
-
-            context 'but does not match version' do
-              let(:package_version) { 'package-version-abcd' }
-
-              it 'copies metadata from the existing package' do
-                expect(update_release).to receive(:create_packages).with(
-                  [
-                    new_package_metadata.merge(
-                      'compiled_package_sha1' => new_package_metadata['sha1'],
-                      'blobstore_id' => blobstore_id,
-                      'sha1' => model_sha,
-                    ),
-                  ],
-                  release_dir,
-                )
-
-                update_release.process_packages(release_dir)
-              end
-            end
-          end
-        end
-      end
-    end
-
     describe 're-using existing packages' do
       subject(:job) { Jobs::UpdateRelease.new(release_path, {}) }
